@jart 2024-04-23
The tinyBLAS code upstreamed by Mozilla's llamafile project makes prompt processing go very fast for F32, F16, Q4\_0, and Q8\_0.

| model                          |       size |     params | backend    |    threads | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ---------- | ---------------: |
| **llama 1B F16**                   |   2.05 GiB |     1.10 B | CPU        |         96 | pp 512     |   **2048.86 ± 6.52** |
| llama 1B F16                   |   2.05 GiB |     1.10 B | CPU        |         96 | tg 4       |     52.01 ± 0.06 |
| **llama 1B all F32**               |   4.10 GiB |     1.10 B | CPU        |         96 | pp 512     |  **1946.19 ± 21.26** |
| llama 1B all F32               |   4.10 GiB |     1.10 B | CPU        |         96 | tg 4       |     39.75 ± 0.15 |
| llama 1B Q2_K - Medium         | 411.41 MiB |     1.10 B | CPU        |         96 | pp 512     |  1273.08 ± 18.57 |
| llama 1B Q2_K - Medium         | 411.41 MiB |     1.10 B | CPU        |         96 | tg 4       |     69.55 ± 0.22 |
| llama 1B Q3_K - Large          | 563.42 MiB |     1.10 B | CPU        |         96 | pp 512     |   1109.08 ± 7.66 |
| llama 1B Q3_K - Large          | 563.42 MiB |     1.10 B | CPU        |         96 | tg 4       |     66.83 ± 0.41 |
| llama 1B Q3_K - Medium         | 522.30 MiB |     1.10 B | CPU        |         96 | pp 512     |   1161.17 ± 7.58 |
| llama 1B Q3_K - Medium         | 522.30 MiB |     1.10 B | CPU        |         96 | tg 4       |     67.49 ± 0.10 |
| llama 1B Q3_K - Small          | 475.51 MiB |     1.10 B | CPU        |         96 | pp 512     | 1052.25 ± 161.64 |
| llama 1B Q3_K - Small          | 475.51 MiB |     1.10 B | CPU        |         96 | tg 4       |     68.30 ± 0.19 |
| **llama 1B Q4_0**                  | 606.53 MiB |     1.10 B | CPU        |         96 | pp 512     |  **1418.41 ± 10.54** |
| llama 1B Q4_0                  | 606.53 MiB |     1.10 B | CPU        |         96 | tg 4       |     65.78 ± 0.24 |
| llama 1B Q4_1                  | 668.18 MiB |     1.10 B | CPU        |         96 | pp 512     |    884.68 ± 3.74 |
| llama 1B Q4_1                  | 668.18 MiB |     1.10 B | CPU        |         96 | tg 4       |     64.62 ± 0.05 |
| llama 1B Q4_K - Medium         | 636.18 MiB |     1.10 B | CPU        |         96 | pp 512     |  1197.76 ± 11.42 |
| llama 1B Q4_K - Medium         | 636.18 MiB |     1.10 B | CPU        |         96 | tg 4       |     66.04 ± 0.34 |
| llama 1B Q4_K - Small          | 609.53 MiB |     1.10 B | CPU        |         96 | pp 512     |  1200.22 ± 10.06 |
| llama 1B Q4_K - Small          | 609.53 MiB |     1.10 B | CPU        |         96 | tg 4       |     66.38 ± 0.27 |
| llama 1B Q5_0                  | 729.84 MiB |     1.10 B | CPU        |         96 | pp 512     |  1058.68 ± 10.52 |
| llama 1B Q5_0                  | 729.84 MiB |     1.10 B | CPU        |         96 | tg 4       |     63.10 ± 0.36 |
| llama 1B Q5_1                  | 791.50 MiB |     1.10 B | CPU        |         96 | pp 512     |  718.18 ± 127.77 |
| llama 1B Q5_1                  | 791.50 MiB |     1.10 B | CPU        |         96 | tg 4       |     62.07 ± 0.67 |
| llama 1B Q5_K - Medium         | 745.11 MiB |     1.10 B | CPU        |         96 | pp 512     |   1055.78 ± 5.80 |
| llama 1B Q5_K - Medium         | 745.11 MiB |     1.10 B | CPU        |         96 | tg 4       |     64.01 ± 0.16 |
| llama 1B Q5_K - Small          | 729.84 MiB |     1.10 B | CPU        |         96 | pp 512     |   1048.20 ± 3.90 |
| llama 1B Q5_K - Small          | 729.84 MiB |     1.10 B | CPU        |         96 | tg 4       |     64.32 ± 0.27 |
| llama 1B Q6_K                  | 860.86 MiB |     1.10 B | CPU        |         96 | pp 512     |  995.96 ± 183.61 |
| llama 1B Q6_K                  | 860.86 MiB |     1.10 B | CPU        |         96 | tg 4       |     62.67 ± 0.22 |
| **llama 1B Q8_0**                  |   1.09 GiB |     1.10 B | CPU        |         96 | pp 512     |   **1430.38 ± 9.86** |
| llama 1B Q8_0                  |   1.09 GiB |     1.10 B | CPU        |         96 | tg 4       |     59.71 ± 0.14 |

Measured on AMD Ryzen Threadripper PRO 7995WX with TinyLlama 1.1B. This PR ensures those performance wins will happen for MoE models too.