@jart 2024-04-24
Oh my gosh, I finally got around to running this code on my Mac Studio M2 Ultra w/ 128 GB RAM for Mixtral 8x7b F16. This change has made prompt processing go 5.5x faster. My mixmul implementation appears to significantly improve token generation speed too.

![image](https://github.com/ggerganov/llama.cpp/assets/49262/e98e3a35-7aab-46c5-9193-7e24f75b40e6)

Here we see that, compared to llama.cpp at HEAD today, prompt tok/sec leaps from 10.66 -> 55.53 and eval tok/sec gets pumped up from 5.58 -> 8.06.

Most MacOS users won't care, since Metal works so well. However this is actually great news for Mac Studio owners who've installed Asahi Linux (which doesn't have Apple Accelerate or Apple Metal). They'll now be able able to run the best models significantly faster, using only the ARMv8.2 ISA and an exclusively free+open+libre software stack.