@jart 2024-04-24
Note: I'm still in the process of testing this change and verifying it's correct on all compilers and architectures.