@jart 2024-04-25
OK I've worked out the remaining kinks. This code was just shipped as part of the [llamafile 0.8 release](https://github.com/Mozilla-Ocho/llamafile/releases/tag/0.8). Thanks to this change, I'm seeing a 2x prompt eval speed increase across the board. My Threadripper now runs Mixtral 2x faster. My M2 Ultra runs Mixtral 2x faster on CPU. This change even pumps up the Raspberry Pi 5 to 78 tok/sec performance on non-MoE F16 models in case you want to buy a bag full of the things to build your next supercomputer. PTAL.