@ikawrakow 2024-04-26
>  It's a shame because I really like Q5_K_M, so it'd be great to see Iwan Kawrakow develop a new quant specifically for block-tiling.

@jart 

I became intrigued by your assumption that block-tiling is required to speed up prompt processing for k-quants, so spent some time optimizing k-quant CPU matrix multiplications. I'm running on a 16-core Ryzen-7950X CPU, so have done just a better AVX2 implementation. Baseline for this CPU (using your PR) for a 7B LLaMA is

| model                          |       size |     params | backend    |    threads | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ---------- | ---------------: |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CPU        |         16 | pp 512     |    119.75 ± 0.38 |
| llama 7B Q4_1                  |   3.95 GiB |     6.74 B | CPU        |         16 | pp 512     |     63.80 ± 0.22 |
| llama 7B Q5_0                  |   4.33 GiB |     6.74 B | CPU        |         16 | pp 512     |     59.50 ± 0.11 |
| llama 7B Q5_1                  |   4.72 GiB |     6.74 B | CPU        |         16 | pp 512     |     56.28 ± 0.08 |

`Q4_0` is much faster than the other legacy quants thanks to your tinyBLAS.

Here is what I get for k-quants

| model                          |       size |     params | test       |    t/s (master) |  t/s (optimized) |  Speedup      |   
| ------------------------------ | ---------: | ---------: | ---------- | ---------------: | ---------------: | ------------: |
| llama 7B Q2_K - Small          |   2.16 GiB |     6.74 B | pp 512     |    116.35 ± 0.08 |    162.66 ± 1.21 | 1.398 ± 0.009 |
| llama 7B Q2_K - Medium         |   2.36 GiB |     6.74 B | pp 512     |    100.49 ± 0.15 |    149.31 ± 0.48 | 1.486 ± 0.006 |
| llama 7B Q3_K - Small          |   2.75 GiB |     6.74 B | pp 512     |     82.01 ± 0.11 |    132.68 ± 0.21 | 1.618 ± 0.003 |
| llama 7B Q3_K - Medium         |   3.07 GiB |     6.74 B | pp 512     |     89.16 ± 0.10 |    136.53 ± 0.19 | 1.531 ± 0.003 |
| llama 7B Q4_K - Small          |   3.59 GiB |     6.74 B | pp 512     |    104.45 ± 0.15 |    144.87 ± 0.31 | 1.387 ± 0.003 |
| llama 7B Q4_K - Medium         |   3.80 GiB |     6.74 B | pp 512     |    101.50 ± 0.28 |    145.77 ± 0.33 | 1.436 ± 0.005 |
| llama 7B Q5_K - Small          |   4.33 GiB |     6.74 B | pp 512     |     73.72 ± 0.29 |    124.05 ± 0.14 | 1.682 ± 0.007 |
| llama 7B Q5_K - Medium         |   4.45 GiB |     6.74 B | pp 512     |     74.84 ± 0.13 |    126.52 ± 0.14 | 1.691 ± 0.004 |
| llama 7B Q6_K                  |   5.15 GiB |     6.74 B | pp 512     |     81.38 ± 0.09 |    146.30 ± 0.15 | 1.798 ± 0.003 |

You favorite `Q5_K_M` quants are faster than `Q4_0` with tinyBLAS with these changes :-)

There are 3 ingredients involved in this speedup:
* Simdify `Q8_K` quantization. Quantization of activations is single-threaded in `ggml`. Quantization to `Q8_0`, needed by legacy quants, is already simdified, but quantization to `Q8_K`, required by k-quants, is not. This does not matter for token generation, but we get a 5-6% speedup for prompt processing.
* Tweak some more the k-quant dot product kernels to reduce/eliminate dependencies between computational steps. I guess this is what you call "instruction level parallelism"
* Last, but certainly not least, make use of the ability to do 2x2 matrix multiplications (rather than just dot products) that is already available in `ggml` (2 weight rows times two activation columns). This gives 20-40% speedup, depending on how costly it is to setup the bits as needed for the multiplication with the `Q8_K` quants. 

We see `Q4_K` and `Q5_K` being ~2.2 times faster than their respective legacy counterparts `Q4_1` and `Q5_1`.