@ggerganov 2024-05-22
On M2 Ultra I'm observing some TG regression for `Q8_0` and `Q4_0` MoE when running fully on the CPU:

```bash
./scripts/compare-commits.sh master pr/6840 \
  -m models/mixtral-instruct-8x7b-fast/ggml-model-f16.gguf \
  -m models/mixtral-instruct-8x7b-fast/ggml-model-q8_0.gguf \
  -m models/mixtral-instruct-8x7b-fast/ggml-model-q4_0.gguf -t 16 -ngl 0
```

| CPU      | Model           |   Model Size [GiB] | Test        |   t/s master |   t/s pr/6840 |   Speedup |
|:---------|:----------------|-------------------:|:------------|-------------:|--------------:|----------:|
| M2 Ultra | llama 8x7B F16  |              86.99 | pp512       |        30.54 |         47.72 |      1.56 |
| M2 Ultra | llama 8x7B F16  |              86.99 | tg128       |         8.75 |          7.66 |      0.88 |
| M2 Ultra | llama 8x7B F16  |              86.99 | pp512+tg128 |        20.42 |         22.85 |      1.12 |
| M2 Ultra | llama 8x7B Q4_0 |              48.25 | pp512       |        44.39 |         63.95 |      1.44 |
| M2 Ultra | llama 8x7B Q4_0 |              48.25 | tg128       |        25.23 |         23.23 |      0.92 |
| M2 Ultra | llama 8x7B Q4_0 |              48.25 | pp512+tg128 |        36.62 |         44.83 |      1.22 |
| M2 Ultra | llama 8x7B Q8_0 |              46.22 | pp512       |        49.20 |         71.55 |      1.45 |
| M2 Ultra | llama 8x7B Q8_0 |              46.22 | tg128       |        15.32 |         15.29 |      1.00 |
| M2 Ultra | llama 8x7B Q8_0 |              46.22 | pp512+tg128 |        32.50 |         39.38 |      1.21 |

@jart Do you observe this regression on your M2 Ultra?